Move harness things to harness/.
